Move faster route out of mapbox navigation
