+++ libnavigation-core/src/main/java/com/mapbox/navigation/core/directions/session/DirectionsSession.kt
@@ -1,14 +1,22 @@
 package com.mapbox.navigation.core.directions.session
 
+import android.location.Location
 import com.mapbox.api.directions.v5.models.DirectionsRoute
 import com.mapbox.api.directions.v5.models.RouteOptions
 import com.mapbox.navigation.base.route.Router
+import com.mapbox.navigation.base.trip.model.RouteProgress
 
 internal interface DirectionsSession {
 
     var routes: List<DirectionsRoute>
 
     fun getRouteOptions(): RouteOptions?
+
+    fun getAdjustedRouteOptions(
+        routeOptions: RouteOptions,
+        routeProgress: RouteProgress,
+        location: Location
+    ): RouteOptions
 
     fun requestRoutes(routeOptions: RouteOptions, routesRequestCallback: RoutesRequestCallback? = null)
